Fix sharing setup
For the --spread option, when assigning threads we first use the
permutation as it is given, and then if we need more permutations we
rotate it cyclically. These rotations are sometimes only done on the
first $n$ elements, where $n$ represents the number of objectives left
to fix.

This means we also need to set up the correct sharing arrays, i.e.
share_to, share_from, share_bounds, share_limit. This patch fixes the
creation of these arrays. index first points to the "last non-fixed"
element of the permutation, but then cycles around to get the first
"nObjLeft - 2" elements as well.