Fix threading
There was an issue where some threads were "waiting to share results but
not finished the whole problem" while others wree "waiting to share
results and the whole problem was solved", which would cause issues.
Mostly with large numbers of threads. This fixes that by using two
distinct locks to differentiate between the two scenarios, and also to
more explicitly track the state of each thread.